Udit-Senapaty_Image-Deepfake-Detection/requirements.txt
torch
torchvision
streamlit
pillow
numpy
